refactor(refactor/tier): 조건 argument resolver 코드도 discovery로 옮김
