+++ src/main/java/com/kustaurant/kustaurant/api/discovery/RestaurantTierApiController.java
@@ -4,9 +4,9 @@ import com.kustaurant.kustaurant.common.restaurant.service.RestaurantApiService;
 import com.kustaurant.kustaurant.common.discovery.service.RestaurantTierService;
 import com.kustaurant.kustaurant.global.UserService;
 import com.kustaurant.kustaurant.global.apiUser.customAnno.JwtToken;
-import com.kustaurant.kustaurant.common.restaurant.argument_resolver.CuisineList;
-import com.kustaurant.kustaurant.common.restaurant.argument_resolver.LocationList;
-import com.kustaurant.kustaurant.common.restaurant.argument_resolver.SituationList;
+import com.kustaurant.kustaurant.common.discovery.argument_resolver.CuisineList;
+import com.kustaurant.kustaurant.common.discovery.argument_resolver.LocationList;
+import com.kustaurant.kustaurant.common.discovery.argument_resolver.SituationList;
 import com.kustaurant.kustaurant.common.discovery.domain.RestaurantTierDTO;
 import com.kustaurant.kustaurant.common.discovery.domain.RestaurantTierMapDTO;
 import com.kustaurant.kustaurant.common.restaurant.service.RestaurantFavoriteService;
