+++ src/main/java/com/kustaurant/kustaurant/common/discovery/argument_resolver/CuisineListArgumentResolver.java
@@ -1,5 +1,6 @@
-package com.kustaurant.kustaurant.common.restaurant.argument_resolver;
+package com.kustaurant.kustaurant.common.discovery.argument_resolver;
 
+import com.kustaurant.kustaurant.common.discovery.enums.CuisineEnum;
 import com.kustaurant.kustaurant.global.exception.exception.ParamException;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.core.MethodParameter;
@@ -12,11 +13,11 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class SituationListArgumentResolver implements HandlerMethodArgumentResolver {
+public class CuisineListArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterAnnotation(SituationList.class) != null && parameter.getParameterType().equals(List.class);
+        return parameter.getParameterAnnotation(CuisineList.class) != null && parameter.getParameterType().equals(List.class);
     }
 
     @Override
@@ -26,26 +27,46 @@ public class SituationListArgumentResolver implements HandlerMethodArgumentResol
                                   WebDataBinderFactory binderFactory) throws Exception {
         boolean isApiRequest = isApiRequest(webRequest);
 
-        String situations = webRequest.getParameter("situations");
-        if (situations == null || situations.isEmpty()) {
+        String cuisines = webRequest.getParameter("cuisines");
+        if (cuisines == null || cuisines.isEmpty()) {
             return null;
+        }
+
+        if (cuisines.contains("ALL") && cuisines.contains(("JH"))) {
+            if (isApiRequest) {
+                // API 요청의 경우 예외를 던집니다.
+                throw new ParamException("cuisines 파라미터 값에 ALL와 JH가 둘 다 있습니다.");
+            } else {
+                // 웹 요청의 경우 클라이언트에 리다이렉트를 지시합니다.
+                HttpServletResponse response = webRequest.getNativeResponse(HttpServletResponse.class);
+                if (response != null) {
+                    response.sendRedirect("/tier"); // 리다이렉트할 URL
+                }
+                return null;
+            }
         }
 
         try {
             // 파라미터가 "ALL"이면 null 반환
-            if (situations.contains("ALL")) {
+            if (cuisines.contains("ALL")) {
                 return null;
             }
 
             // 문자열을 List<String>으로 변환
-            return Arrays.stream(situations.split(","))
-                    .map(c -> Integer.parseInt(c.trim()))
+            List<String> cuisineList = Arrays.stream(cuisines.split(","))
+                    .map(c -> CuisineEnum.valueOf(c.trim()).getValue())
                     .collect(Collectors.toList());
+
+            if (cuisineList.contains("제휴업체")) {
+                cuisineList = List.of("JH");
+            }
+
+            return cuisineList;
 
         } catch (IllegalArgumentException e) {
             if (isApiRequest) {
                 // API 요청의 경우 예외를 던집니다.
-                throw new ParamException("situations 파라미터 입력이 올바르지 않습니다.");
+                throw new ParamException("cuisines 파라미터 값이 올바르지 않습니다.");
             } else {
                 // 웹 요청의 경우 클라이언트에 리다이렉트를 지시합니다.
                 HttpServletResponse response = webRequest.getNativeResponse(HttpServletResponse.class);
