+++ src/main/java/com/kustaurant/kustaurant/common/discovery/argument_resolver/LocationList.java
@@ -1,4 +1,4 @@
-package com.kustaurant.kustaurant.common.restaurant.argument_resolver;
+package com.kustaurant.kustaurant.common.discovery.argument_resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
