+++ src/main/java/com/kustaurant/kustaurant/common/discovery/argument_resolver/LocationListArgumentResolver.java
@@ -1,6 +1,6 @@
-package com.kustaurant.kustaurant.common.restaurant.argument_resolver;
+package com.kustaurant.kustaurant.common.discovery.argument_resolver;
 
-import com.kustaurant.kustaurant.common.discovery.enums.CuisineEnum;
+import com.kustaurant.kustaurant.common.discovery.enums.LocationEnum;
 import com.kustaurant.kustaurant.global.exception.exception.ParamException;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.core.MethodParameter;
@@ -13,11 +13,11 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class CuisineListArgumentResolver implements HandlerMethodArgumentResolver {
+public class LocationListArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterAnnotation(CuisineList.class) != null && parameter.getParameterType().equals(List.class);
+        return parameter.getParameterAnnotation(LocationList.class) != null && parameter.getParameterType().equals(List.class);
     }
 
     @Override
@@ -27,46 +27,26 @@ public class CuisineListArgumentResolver implements HandlerMethodArgumentResolve
                                   WebDataBinderFactory binderFactory) throws Exception {
         boolean isApiRequest = isApiRequest(webRequest);
 
-        String cuisines = webRequest.getParameter("cuisines");
-        if (cuisines == null || cuisines.isEmpty()) {
+        String locations = webRequest.getParameter("locations");
+        if (locations == null || locations.isEmpty()) {
             return null;
-        }
-
-        if (cuisines.contains("ALL") && cuisines.contains(("JH"))) {
-            if (isApiRequest) {
-                // API 요청의 경우 예외를 던집니다.
-                throw new ParamException("cuisines 파라미터 값에 ALL와 JH가 둘 다 있습니다.");
-            } else {
-                // 웹 요청의 경우 클라이언트에 리다이렉트를 지시합니다.
-                HttpServletResponse response = webRequest.getNativeResponse(HttpServletResponse.class);
-                if (response != null) {
-                    response.sendRedirect("/tier"); // 리다이렉트할 URL
-                }
-                return null;
-            }
         }
 
         try {
             // 파라미터가 "ALL"이면 null 반환
-            if (cuisines.contains("ALL")) {
+            if (locations.contains("ALL")) {
                 return null;
             }
 
             // 문자열을 List<String>으로 변환
-            List<String> cuisineList = Arrays.stream(cuisines.split(","))
-                    .map(c -> CuisineEnum.valueOf(c.trim()).getValue())
+            return Arrays.stream(locations.split(","))
+                    .map(c -> LocationEnum.valueOf(c.trim()).getValue())
                     .collect(Collectors.toList());
-
-            if (cuisineList.contains("제휴업체")) {
-                cuisineList = List.of("JH");
-            }
-
-            return cuisineList;
 
         } catch (IllegalArgumentException e) {
             if (isApiRequest) {
                 // API 요청의 경우 예외를 던집니다.
-                throw new ParamException("cuisines 파라미터 값이 올바르지 않습니다.");
+                throw new ParamException("locations 파라미터 입력이 올바르지 않습니다.");
             } else {
                 // 웹 요청의 경우 클라이언트에 리다이렉트를 지시합니다.
                 HttpServletResponse response = webRequest.getNativeResponse(HttpServletResponse.class);
